backup/requirements.txt
# ModernTensor Core SDK (local development)
-e ../moderntensor_aptos

# Core Python libraries
numpy>=1.24.0
pandas>=2.0.0

# Image processing and AI
Pillow>=10.0.0
diffusers>=0.21.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Async HTTP client
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
cryptography>=41.0.0
httpx>=0.24.0
rich>=13.0.0

# Configuration and Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
coloredlogs>=15.0.0
structlog>=23.0.0

# System monitoring
psutil>=5.9.0
prometheus_client>=0.17.0

# Data validation and serialization
marshmallow>=3.19.0

# Blockchain - Core Blockchain (replaces aptos-sdk)
web3>=6.0.0
eth-account>=0.9.0
eth-keys>=0.4.0
eth-utils>=2.0.0
eth-abi>=4.0.0

# Cryptography for Core blockchain
pynacl>=1.5.0
coincurve>=18.0.0

# AI/ML dependencies for image scoring
clip-by-openai>=1.0.1
scikit-learn>=1.3.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Performance optimization
orjson>=3.9.0  # Fast JSON parsing